order list of files in text file by compilation priority
Older JULES makefiles used to have a specific order in which to compile the files, in order to avoid missing dependencies when a file is being compiled. This order is as follows:

SRC_DIRS = SOURCE/MODULES/PARAMS \
           SOURCE/IMOGEN/MODULES/PARAMS \
           SOURCE/IMOGEN/MODULES/VARS \
           SOURCE/MODULES/CONTROL \
           SOURCE/IMOGEN/MODULES/MAP \
           SOURCE/SUBROUTINES/VEGETATION \
           SOURCE/SUBROUTINES/CONTROL \
           SOURCE/SUBROUTINES/INITIALISATION \
           SOURCE/SUBROUTINES/RADIATION \
           SOURCE/SUBROUTINES/SNOW \
           SOURCE/SUBROUTINES/SOIL \
           SOURCE/SUBROUTINES/SURFACE \
           SOURCE/IMOGEN/SUBROUTINES

This commit makes use of this information to list the files required by decreasing order of compilation of priority (first item has highest priority). Within a given priority level, there is an arbitrary alphabetical ordering (which avoids a problem with um_types and precision_mod - but this may create other problems elsewhere so this may be reviewed later).