fix problem when other_child not added if variable
for other children in a given parent (i.e. module). they were only added to next_callers if they were callers themselves, which meant that they weren't created as a node if they were a variable, and the parenthood edge was also missing. This is now fixed by removing the check on whether child is in caller_callees.